feat: adicionando funções para excluir e editar chamados com validação

+++ manutencao/src/main/java/br/com/matheus/manutencao/dto/LoginRequestDTO.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class LoginRequestDTO {
 
-    private Long matricula;
+    private Integer matricula;
 
     private String nome;
 }
